Figuring out how to make case insensitive properties for Powershell.

+++ AutomationCore/Automation.Core.Tests/DynamicTests.cs
@@ -45,7 +45,7 @@ namespace Automation.Core.Tests
             Assert.IsNull(tree.Parent);
 
             // Case insensitive
-            Assert.AreEqual(anyType, tree.type);
+            Assert.AreEqual(anyType, (int)tree.type);
             Assert.AreSame(child, tree.CHILDREN[0]);
         }
